requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
dnspython==2.6.1
email-validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
msgpack==1.0.8
multidict==6.0.5
nonebot-adapter-onebot==2.4.3
nonebot2==2.3.0
numpy==1.26.4
opencv-python==4.9.0.80
orjson==3.10.3
pydantic==2.7.1
pydantic-core==2.18.2
pygments==2.18.0
pygtrie==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.9
pyyaml==6.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing-extensions==4.11.0
ujson==5.9.0
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.4